Graph Based Parser
Switched everything to a graph based parser instead of a recursive
parser that depends on ordering.

The graph based parser allows us to detect circular dependencies, and
does not enforce any ordering in the config file.

It works by building a directed graph with a node for each key-value pair in the
config. Any references that ocurr in the value of a node (things like R.test_1)
create a dependency of that node on the referenced node. This
dependency takes the form of a directed edge point from the node
containing the reference to the node that is referred to.

We build this directed acyclic graph by recursing through the config
once, constructing edges along the way. We then do a topological sort on
the resulting graph, which tells us the order in which we need to parse
nodes such that all nodes containing dependencies will have their
dependencies properly parsed before being parsed themselves.

Pretty cool! It still fails tests with configs containing update and
extend functions at the same level. This is because correct parsing
of those configs relies on the extend function getting called before the
update function, or vice versa. No dependencies are created between
functions yet, but a quick, hacky solution would be to create special
_depends key that enforces a dependency on some other node in the
config.